first pass refactor
