+++ Makefile
@@ -11,3 +11,7 @@ build-docker:
 		--squash .
 	docker image save serum-crank:latest -o serum_crank.tar
 	pigz -f -9 serum_crank.tar
+
+.PHONY: fmt
+fmt:
+	find -type f -name "*.rs" -not -path "*target*" -exec rustfmt --edition 2018 {} \;
